backend/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.8.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.3.0
pytest==9.0.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
